detour: move littleEndian serialization into own file
